+++ resources/views/Admin/Alumni/index.blade.php
@@ -43,7 +43,7 @@
                                   <td>{{ $item->kontak }}</td>
                                   <td>
                                       <a href="alumni/detail/{{ $item->nisn }}"><button type="button" class="btn btn-icon btn-outline-primary btn-sm"><i class="bx bx-detail"></button></i></a>
-                                      <a href="alumni/delete/{{ $item->nisn }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm"><i class="bx bx-trash-alt"></button></i></a>
+                                      <a href="alumni/delete/{{ $item->nisn }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" onclick="return-window.confirm('Hapus Data Ini?')"><i class="bx bx-trash-alt"></button></i></a>
                                   </td>
                                 </tr>
                                 @endforeach
