+++ resources/views/Admin/Contact/index.blade.php
@@ -24,7 +24,7 @@
                                 <td>{{ $item->email }}</td>
                                 <td>{{ $item->message }}</td>
                                 <td>
-                                    <a href="contact/delete/{{ $item->id_contact }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" ><i class="bx bx-trash-alt"></button></i></a>
+                                    <a href="contact/delete/{{ $item->id_contact }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" onclick="return-window.confirm('Hapus Data Ini?')"><i class="bx bx-trash-alt"></button></i></a>
                                 </td>
                             </tr>
                             @endforeach
