+++ resources/views/Admin/Gallery/index.blade.php
@@ -25,7 +25,7 @@
                                 <th>Aksi</th>
                             </tr>
                         </thead>
-                        <tbody class="table table-bordered text-center">
+                        <tbody class="table-bordered text-center">
                             @foreach ($gallery as $key => $item)
                             <tr>
                                 <td>{{ $key+=1 }}</td>
@@ -34,7 +34,7 @@
                                 </td>
                                 <td>
                                     <a href="gallery/edit/{{ $item->id_gallery }}"><button type="button" class="btn btn-icon btn-outline-primary btn-sm"><i class="bx bx-edit-alt me-1"></button></i></a>
-                                    <a href="gallery/delete/{{ $item->id_gallery }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" ><i class="bx bx-trash-alt"></button></i></a>
+                                    <a href="gallery/delete/{{ $item->id_gallery }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" onclick="return-window.confirm('Hapus Data Ini?')"><i class="bx bx-trash-alt"></button></i></a>
                                 </td>
                             </tr>
                             @endforeach
