+++ resources/views/Admin/Lamaran/index.blade.php
@@ -40,7 +40,7 @@
                                             <td>
                                                 <a href="lamaran/detail/{{ $item->id_lamaran }}"><button type="button" class="btn btn-icon btn-outline-primary btn-sm"><i class="bx bx-detail"></button></i></a>
                                                 <a href="lamaran/edit/{{ $item->id_lamaran }}"><button type="button" class="btn btn-icon btn-outline-primary btn-sm"><i class="bx bx-edit-alt me-1"></button></i></a>
-                                                <a href="lamaran/delete/{{ $item->id_lamaran }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" ><i class="bx bx-trash-alt"></button></i></a>
+                                                <a href="lamaran/delete/{{ $item->id_lamaran }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" onclick="return-window.confirm('Hapus Data Ini?')"><i class="bx bx-trash-alt"></button></i></a>
                                             </td>
                                         </tr>
                                     @endforeach
