+++ resources/views/Admin/Lowongan/index.blade.php
@@ -29,7 +29,7 @@
                                     <th>Aksi</th>
                                 </tr>
                             </thead>
-                            <tbody class="table table-bordered text-center">
+                            <tbody class="table-bordered text-center">
                                 @foreach ($lowongan as $key => $item)
                                 <tr>
                                     <td>{{ $key+=1 }}</td>
@@ -40,8 +40,8 @@
                                     <td>{{ $item->gaji }}</td>
                                     <td>
                                         <a href="lowongan/detail/{{ $item->id_lowongan }}"><button type="button" class="btn btn-icon btn-outline-primary btn-sm mb-2"><i class="bx bx-detail"></button></i></a>
-                                        <a href="lowongan/edit/{{ $item->id_lowongan }}"><button type="button" class="btn btn-icon btn-outline-primary btn-sm mb-2"><i class="bx bx-edit-alt me-1"></button></i></a>
-                                        <a href="lowongan/delete/{{ $item->id_lowongan }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" ><i class="bx bx-trash-alt"></button></i></a>
+                                        <a href="lowongan/edit/{{ $item->id_lowongan }}"><button type="button" class="btn btn-icon btn-outline-primary btn-sm mb-2"><i class="bx bx-edit-alt"></button></i></a>
+                                        <a href="lowongan/delete/{{ $item->id_lowongan }}"><button type="button" class="btn btn-icon btn-outline-danger btn-sm" onclick="return-window.confirm('Hapus Data Ini?')"><i class="bx bx-trash-alt"></button></i></a>
                                     </td>
                                 </tr>
                                 @endforeach
